requirements.txt
pyboy>=2.6.0
openai>=1.0.0
azure-identity>=1.15.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
